requirements.txt
requests==2.31.0
PyMuPDF==1.23.8
Pillow==10.2.0
reportlab==4.0.8 